+++ resources/assets/sass/app.scss
@@ -7,3 +7,9 @@
 
 // Bootstrap
 @import "~bootstrap-sass/assets/stylesheets/bootstrap";
+
+
+#map {
+    width: 100%;
+    height: 100%;
+}
