+++ webpack.mix.js
@@ -11,5 +11,9 @@ let mix = require('laravel-mix');
  |
  */
 
-mix.js('resources/assets/js/app.js', 'public/js')
-   .sass('resources/assets/sass/app.scss', 'public/css');
+mix.js([
+    'resources/assets/js/app.js',
+    'resources/assets/js/script.js',
+    'resources/assets/js/bootstrap.min.js'
+], 'public/js')
+.sass('resources/assets/sass/app.scss', 'public/css');
